Refactor web socket connection stuff
So that it's not duplicated across both screens but handled in the main
App component as much as possible
